Split commands_test.go.
