+++ apiclients/bible/bible_test.go
@@ -47,7 +47,7 @@ func TestFetchUser(t *testing.T) {
 	}
 
 	for _, tc := range tests {
-		server.Resp = tc.useResp
+		server.Resps = []string{tc.useResp}
 		t.Run(tc.desc, func(t *testing.T) {
 			got, err := FetchVerses("Philippians 4:8")
 			if err != nil {
