+++ apiclients/kick/kick_test.go
@@ -610,7 +610,7 @@ func TestFetchChannel(t *testing.T) {
 	}
 
 	for _, tc := range tests {
-		server.Resp = tc.useResp
+		server.Resps = []string{tc.useResp}
 		t.Run(tc.desc, func(t *testing.T) {
 			got, err := kick.FetchChannel("user1")
 			if err != nil {
