+++ commands/moderation/moderation.go
@@ -22,9 +22,9 @@ var (
 		Desc:       "Times you out for 1 second.",
 		Permission: permission.Normal,
 		Handler: func(msg *base.IncomingMessage, args []arg.Arg) ([]*base.Message, error) {
-			err := msg.Platform.Timeout(msg.Message.User, msg.Message.Channel, 1*time.Second)
+			err := msg.Resources.Platform.Timeout(msg.Message.User, msg.Message.Channel, 1*time.Second)
 			if err != nil {
-				return nil, fmt.Errorf("failed to time out %s/%s on %s: %w", msg.Message.Channel, msg.Message.User, msg.Platform.Name(), err)
+				return nil, fmt.Errorf("failed to time out %s/%s on %s: %w", msg.Message.Channel, msg.Message.User, msg.Resources.Platform.Name(), err)
 			}
 			return nil, nil
 		},
