+++ main.go
@@ -25,29 +25,34 @@ const (
 	waitForContextCancellation = 100 * time.Millisecond
 )
 
+type postStartupResources struct {
+	platforms map[string]base.Platform
+	cache     cache.Cache
+}
+
 func initialStart(ctx context.Context) (cleanup.Cleaner, error) {
 	cleaner, _, err := start(ctx)
 	return cleaner, err
 }
 
 func reStart(ctx context.Context) (cleanup.Cleaner, error) {
-	cleaner, ps, err := start(ctx)
+	cleaner, resources, err := start(ctx)
 	if err != nil {
 		return nil, err
 	}
-	if err := restart.Notify(ps); err != nil {
+	if err := restart.Notify(resources.cache, resources.platforms); err != nil {
 		log.Printf("Failed to notify restart: %v", err)
 	}
 	return cleaner, err
 }
 
-func start(ctx context.Context) (cleanup.Cleaner, map[string]base.Platform, error) {
+func start(ctx context.Context) (cleanup.Cleaner, postStartupResources, error) {
 	cleaner := cleanup.NewCleaner()
 
 	log.Print("Reading config...")
 	cfg, err := config.Read()
 	if err != nil {
-		return nil, nil, fmt.Errorf("failed to read config: %v", err)
+		return nil, postStartupResources{}, fmt.Errorf("failed to read config: %v", err)
 	}
 
 	log.Print("Setting config values...")
@@ -56,29 +61,27 @@ func start(ctx context.Context) (cleanup.Cleaner, map[string]base.Platform, erro
 	log.Printf("Connecting to database...")
 	db, err := database.Connect(ctx, os.Getenv("POSTGRES_DB"), os.Getenv("POSTGRES_USER"), os.Getenv("POSTGRES_PASSWORD"))
 	if err != nil {
-		return nil, nil, fmt.Errorf("failed to connect to database: %v", err)
+		return nil, postStartupResources{}, fmt.Errorf("failed to connect to database: %v", err)
 	}
-	database.SetInstance(db)
 
 	log.Printf("Connecting to cache...")
 	cdb := cache.NewRedis()
-	cache.SetInstance(&cdb)
 
 	log.Printf("Performing database migrations...")
 	if err = database.Migrate(db); err != nil {
-		return nil, nil, fmt.Errorf("failed to perform database migrations: %v", err)
+		return nil, postStartupResources{}, fmt.Errorf("failed to perform database migrations: %v", err)
 	}
 
 	log.Printf("Preparing chat connections...")
 	ps, err := platforms.Build(cfg, db, &cdb)
 	if err != nil {
-		return nil, nil, fmt.Errorf("failed to build platforms: %v", err)
+		return nil, postStartupResources{}, fmt.Errorf("failed to build platforms: %v", err)
 	}
 
 	for _, p := range ps {
 		log.Printf("Connecting to %s...", p.Name())
 		if err := p.Connect(); err != nil {
-			return cleaner, nil, fmt.Errorf("failed to connect to %s: %v", p.Name(), err)
+			return cleaner, postStartupResources{}, fmt.Errorf("failed to connect to %s: %v", p.Name(), err)
 		}
 
 		log.Printf("Starting to handle messages on %s...", p.Name())
@@ -94,7 +97,7 @@ func start(ctx context.Context) (cleanup.Cleaner, map[string]base.Platform, erro
 		go supinicClient.StartPinging(ctx)
 	}
 
-	return cleaner, ps, nil
+	return cleaner, postStartupResources{cache: &cdb, platforms: ps}, nil
 }
 
 func main() {
