+++ platforms/platforms.go
@@ -3,6 +3,7 @@ package platforms
 
 import (
 	"context"
+	"crypto/rand"
 	"log"
 	"runtime/debug"
 	"time"
@@ -33,7 +34,7 @@ func Build(cfg *config.Config, db *gorm.DB, cdb cache.Cache) (map[string]base.Pl
 // StartHandling starts handling commands coming from the given platform.
 // This function blocks and should be run within a goroutine.
 func StartHandling(ctx context.Context, p base.Platform, db *gorm.DB, cdb cache.Cache, logIncoming, logOutgoing bool) {
-	handler := commands.NewHandler(db)
+	handler := commands.NewHandler(db, cdb, base.RandResources{Reader: rand.Reader})
 	inC := p.Listen()
 
 	outC := make(chan base.OutgoingMessage, 100)
@@ -114,7 +115,7 @@ func startSending(ctx context.Context, p base.Platform, outC <-chan base.Outgoin
 				}
 			}
 
-			slowmode, err := cdb.FetchBool(cache.GlobalSlowmodeKey(p))
+			slowmode, err := cdb.FetchBool(cache.GlobalSlowmodeKey(p.Name()))
 			if err != nil {
 				log.Printf("Failed to fetch slowmode status for %s: %v", p.Name(), err)
 			}
